# backend/docs/swagger-output.json
{
  "openapi": "2.0",
  "info": {
    "title": "GLY VTU API",
    "description": "Auto-generated Swagger specification for GLY VTU backend.",
    "version": "3.4.4"
  },
  "servers": [
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/overview": {
      "get": {
        "tags": [
          "Admin Finance"
        ],
        "summary": "Finance overview metrics",
        "description": "",
        "responses": {
          "200": {
            "description": "Overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceOverview"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceOverview"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/balances": {
      "get": {
        "tags": [
          "Admin Finance"
        ],
        "summary": "List user wallet balances",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Balances",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/WalletBalanceRow"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/WalletBalanceRow"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/export": {
      "get": {
        "tags": [
          "Admin Finance"
        ],
        "summary": "Export finance report",
        "description": "",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "description": "csv or pdf",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV or PDF report"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/adjustments": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/adjustments/{id}/approve": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-idempotency-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/adjustments/{id}/reject": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user-agent",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/virtual-accounts/requery": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FinanceOverview": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "users": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "volume": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "revenue": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "credits": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "debits": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "walletBalance": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "users": {
                "type": "number",
                "example": 120
              },
              "volume": {
                "type": "number",
                "example": 500000
              },
              "revenue": {
                "type": "number",
                "example": 12000
              },
              "credits": {
                "type": "number",
                "example": 250000
              },
              "debits": {
                "type": "number",
                "example": 230000
              },
              "walletBalance": {
                "type": "number",
                "example": 45000
              }
            }
          }
        },
        "xml": {
          "name": "FinanceOverview"
        }
      },
      "WalletBalanceRow": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "full_name": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "email": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "balance": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "currency": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "updated_at": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          },
          "example": {
            "type": "object",
            "properties": {
              "full_name": {
                "type": "string",
                "example": "Ada Lovelace"
              },
              "email": {
                "type": "string",
                "example": "ada@example.com"
              },
              "balance": {
                "type": "number",
                "example": 1200
              },
              "currency": {
                "type": "string",
                "example": "NGN"
              },
              "updated_at": {
                "type": "string",
                "example": "2026-03-27T10:00:00Z"
              }
            }
          }
        },
        "xml": {
          "name": "WalletBalanceRow"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Use: `Bearer <token>`"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}